+++ auth/client/src/main/java/org/wildfly/security/auth/client/MatchRule.java
@@ -18,7 +18,7 @@
 
 package org.wildfly.security.auth.client;
 
-import static org.wildfly.security.auth.client.ElytronMessages.log;
+import static org.wildfly.security.auth.client._private.ElytronMessages.log;
 
 import java.net.Inet6Address;
 import java.net.InetAddress;
