+++ auth/client/src/main/java/org/wildfly/security/auth/client/_private/ElytronMessages.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.wildfly.security.auth.client;
+package org.wildfly.security.auth.client._private;
 
 import static org.jboss.logging.Logger.Level.WARN;
 
@@ -68,7 +68,7 @@ import org.wildfly.client.config.XMLLocation;
         @ValidIdRange(min = 9529, max = 9529),
         @ValidIdRange(min = 14000, max = 14999)
 })
-interface ElytronMessages extends BasicLogger {
+public interface ElytronMessages extends BasicLogger {
 
     ElytronMessages log = Logger.getMessageLogger(ElytronMessages.class, "org.wildfly.security");
     ElytronMessages xmlLog = Logger.getMessageLogger(ElytronMessages.class, "org.wildfly.security.xml");
