+++ wildfly-elytron/pom.xml
@@ -924,6 +924,7 @@
                                     <include>org.wildfly.security.x500.cert</include>
                                 </includes>
                                 <excludes>
+                                    <exclude>org.wildfly.security.auth.client.ElytronMessages_$logger</exclude>
                                 </excludes>
                                 <includeExclusively>true</includeExclusively>
                             </parameter>
